+++ courier-java-core/src/main/kotlin/com/courier/services/blocking/notifications/CheckService.kt
@@ -28,6 +28,7 @@ interface CheckService {
      */
     fun withOptions(modifier: Consumer<ClientOptions.Builder>): CheckService
 
+    /** Replace the checks for a notification template submission. */
     fun update(submissionId: String, params: CheckUpdateParams): CheckUpdateResponse =
         update(submissionId, params, RequestOptions.none())
 
@@ -49,6 +50,7 @@ interface CheckService {
         requestOptions: RequestOptions = RequestOptions.none(),
     ): CheckUpdateResponse
 
+    /** Retrieve the checks for a notification template submission. */
     fun list(submissionId: String, params: CheckListParams): CheckListResponse =
         list(submissionId, params, RequestOptions.none())
 
@@ -69,6 +71,7 @@ interface CheckService {
         requestOptions: RequestOptions = RequestOptions.none(),
     ): CheckListResponse
 
+    /** Cancel a notification template submission. */
     fun delete(submissionId: String, params: CheckDeleteParams) =
         delete(submissionId, params, RequestOptions.none())